rename folder and file names
